Got most of the music parts working!
Basically after the disaster that was trying to get discord-player to work I decided to implement most of checks myself and the general logic that the dependency used to have. I don't know if somebody took over Androz or something like that but this system is really unintuitive and missing key features that made the original so friendly to use. Anyhow that's my little rant, now most of it is working some fixes need to be made.

Crucial util was added the two way map! Yes without the goddamn maps because for some reason my nodejs Map objects are just broken -w-. Yes anyhow that is here and ready to do it's job