Major update
Basically due to Discord's efforts being to discontinue message based bots, instead forcing devs to use interactions, I had to remake a lot.

Basically biggest changes in order:
-Slash Command Structure
-Interaction event with proper handling
-Linter is in placed and everything had it's consistency increased
-Some utility additions (notable is the one that takes all of the options out of an interaction)

Now I think I will direct all of the focus onto the website and the database, when that's in place the other commands are either easy or are rather plug and play. Because quite frankly everything up to this point had been laying the groundwork for a modular app. Now I just need to do the same for the web and db.