Serverside work, frontend rollback
So mostly getting authentification ready for server, also some other functions. But mainly focusing on MongoDB and Express.js stuff for now so that I can get users authentificated easily from that point it's just fetching from frontend to server to Discord bot client. Nothing too difficult.

Frontend sadly had to be deleted and recreated with CRA. Pretty sad I couldn't use RMW but it lacks some pretty important features I need. Issue isn't that I can't navigate the package but the fact that it relies on custom modules which I can't easily modify and I am not in mood to fork the project and rewrite it just to suit my needs right now (maybe once I have more time). So for now it's pretty much reset to point zero but should be relatively fine for the things I want to use.