Frontend rough work finished
Frontend is fully functional, it looks well... ass. But it's fine for now. I can always spend entire afternoon trying to juggle MUI system and CSS properties, but it was more important for it to work. And it was for sure difficult therefore a lot of the code might look very unreadable. That will improve and it will be documented, but honestly no one should be snooping there anyway since it's not relevant

Backend is done in terms of server for now, more to come as I enable other features I have planned. For now focus back on bot for it's commands